\name{sociosexuality}
\alias{sociosexuality}

\docType{data}

\title{Example data on sociosexual orientations}

\description{
This data set contains simulated data on demographic variables (sex, preferred sex in romantic relationships, education, age) and sociosexual behavior, attitudes, and desire.

The data have no inherent meaning and are intended only for illustration purposes. The variable names were chosen to mimic those in the data from Jones and DeBruine (2019).
}

\format{Data frame with 500 observations on 7 variables.}

\usage{data(sociosexuality)}

\references{
Jones, B. C., & DeBruine, L. (2019). Sociosexuality and self-rated attractiveness. \doi{10.17605/osf.io/6bk3w}
}

\keyword{datasets}
